Applied prettier and fixed solhint warnings
Applied prettier command after changing settings (copied from AAVE but
with a few changes) and fixed other solhint errors (mostly mixed case)

+++ contracts/mocks/TestCurrency.sol
@@ -7,19 +7,22 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 contract TestCurrency is ERC20 {
   address private _owner;
 
-  constructor(string memory name_, string memory symbol_, uint256 initialSupply) ERC20(name_, symbol_) {
+  constructor(
+    string memory name_,
+    string memory symbol_,
+    uint256 initialSupply
+  ) ERC20(name_, symbol_) {
     _owner = msg.sender;
     _mint(msg.sender, initialSupply);
   }
 
-  function mint(address recipient, uint amount) public {
+  function mint(address recipient, uint256 amount) public {
     // require(msg.sender == _owner, "Only owner can mint");
     return _mint(recipient, amount);
   }
 
-  function burn(address recipient, uint amount) public {
+  function burn(address recipient, uint256 amount) public {
     // require(msg.sender == _owner, "Only owner can burn");
     return _burn(recipient, amount);
   }
-
 }
